api/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
supabase
gotrue
boto3
google-generativeai
python-jose[cryptography]
python-multipart